(Fix win-x64) Change naming of SwiftShader libs.
A problem was discovered on Windows x64
where there was a black screen on startup.

It was determined that the incorrect libraries
were being linked.

Then, a feature was found in the emulator
where it would look specifically for
libEGL/libGLES_CM/libGLESv2.

SwiftShader prebuilts will now use this feature.

Note: These will all appropriately display
the SwiftShader watermark.

Change-Id: I62fdb25e0b7a0c3bfb965eb24535e4aef0355954